--- a/01-go-level-1/fibonacci/fiblib/benchmark_test.go
+++ b/01-go-level-1/fibonacci/fiblib/benchmark_test.go
@@ -1,0 +1,40 @@
+package fiblib
+
+import "testing"
+
+var (
+	GlobalRes uint64
+	GlobalN   uint64 = 10
+)
+
+func BenchmarkFibonacciRecur(b *testing.B) {
+	var res uint64
+	for i := 0; i < b.N; i++ {
+		res = FibonacciRecur(GlobalN)
+	}
+	GlobalRes = res
+}
+
+func BenchmarkFibonacciGoldenRatio(b *testing.B) {
+	var res uint64
+	for i := 0; i < b.N; i++ {
+		res = FibonacciRecur(GlobalN)
+	}
+	GlobalRes = res
+}
+
+func BenchmarkFibonacciMapUse(b *testing.B) {
+	var res uint64
+	for i := 0; i < b.N; i++ {
+		res = FibonacciRecur(GlobalN)
+	}
+	GlobalRes = res
+}
+
+func BenchmarkFibonacciAnonymousFunc(b *testing.B) {
+	var res uint64
+	for i := 0; i < b.N; i++ {
+		res = FibonacciRecur(GlobalN)
+	}
+	GlobalRes = res
+}