--- a/01-go-level-1/fibonacci/fiblib/example_test.go
+++ b/01-go-level-1/fibonacci/fiblib/example_test.go
@@ -1,0 +1,8 @@
+package fiblib
+
+import "fmt"
+
+func ExampleFibonacciGoldenRatio() {
+	fmt.Printf("%d", FibonacciGoldenRatio(3))
+	// Output: 2
+}