--- a/01-go-level-1/fibonacci/fiblib/fib_test.go
+++ b/01-go-level-1/fibonacci/fiblib/fib_test.go
@@ -1,0 +1,13 @@
+package fiblib
+
+import "testing"
+
+var vf = []uint64{0, 1, 1, 2, 3, 5, 8, 13, 21, 34, 55, 89, 144, 233, 377, 610, 987, 1597, 2584, 4181, 6765}
+
+func TestFibonacciMapUse(t *testing.T) {
+	for i := 0; i < len(vf); i++ {
+		if f := FibonacciMapUse(uint64(i)); f != vf[i] {
+			t.Logf("Fibonacci %d = %d, want - %d", i, f, vf[i])
+		}
+	}
+}