--- a/01-go-level-1/fibonacci/fiblib/tables_test.go
+++ b/01-go-level-1/fibonacci/fiblib/tables_test.go
@@ -1,0 +1,24 @@
+package fiblib
+
+import "testing"
+
+func TestFibonacciGoldenRatio(t *testing.T) {
+	tests := []struct {
+		name string
+		in   uint64
+		want uint64
+	}{
+		{"simple", 3, 2},
+		{"negative", 0, 100},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			t.Parallel()
+			got := FibonacciGoldenRatio(tc.in)
+			if got != tc.want {
+				t.Fatalf("%s: expected %v; got - %v", tc.name, tc.want, got)
+			}
+		})
+	}
+}