1) Добавлены тесты и бенчмарки для Фибоначчи
2) Бенчмарки для разных реализаций сортировки. Из бенчмарков видно, насколько эффективнее алгоритм быстрой сортировки (по количеству операций):
BenchmarkInsertationSort-4      12012099                96.8 ns/op
BenchmarkBubleSort-4                 7925994                 140 ns/op
BenchmarkQuickSort-4                   357414               4218 ns/op

+ benchstat в формате csv:
benchstat -csv /tmp/quickSort.txt /tmp/insertSort.txt
name,old time/op (ns/op),±,new time/op (ns/op),±,delta,±
Sort-4,3.08500E+03,0%,1.42000E+02,0%,~,(p=1.000 n=1+1)